Haciendo el hook de compra Tiquete
